backend/requirements.txt
Django>=4.2,<5.0
djangorestframework
psycopg2-binary
djangorestframework-simplejwt==5.3.1
djangorestframework-simplejwt==5.3.1
cryptography==42.0.5